Extract the log-likelihood of an estimated model

The loglik_array() methods for measrdcm objects calculates the log-likelihood for an estimated model via the generated quantities functionality in Stan and returns the draws of the log_lik parameter.

Usage

loglik_array(model, ...)

Arguments

model

A measrdcm object.

...

Unused. For future extensions.

Value

A "draws_array" object containing the log-likelihood estimates for the model.