measr: Bayesian Psychometric Measurement Using 'Stan'

Estimate diagnostic classification models (also called cognitive diagnostic models) with 'Stan'. Diagnostic classification models are confirmatory latent class models, as described by Rupp et al. (2010, ISBN: 978-1-60623-527-0). Automatically generate 'Stan' code for the general loglinear cognitive diagnostic diagnostic model proposed by Henson et al. (2009) doi:10.1007/s11336-008-9089-5 and other subtypes that introduce additional model constraints. Using the generated 'Stan' code, estimate the model evaluate the model's performance using model fit indices, information criteria, and reliability metrics.

See also

Author

Maintainer: W. Jake Thompson wjakethompson@gmail.com (ORCID)

Other contributors:

Nathan Jones jonesnateb@gmail.com (ORCID) [contributor]

Matthew Johnson (Provided code adapted for reliability.measrdcm()) [copyright holder]

Paul-Christian Bürkner (Author of eval_silent()) [copyright holder]

University of Kansas [copyright holder]

Institute of Education Sciences [funder]