Extract components of a measrfit object.

Extract components of a measrfit object.

Extract components of an estimated diagnostic classification model

Usage

measr_extract(model, ...)

# S3 method for measrdcm
measr_extract(model, what, ppmc_interval = 0.95, quiet = FALSE, ...)

Arguments

model

The estimated to extract information from.

...

Additional arguments passed to specific methods.

what

Character string. The information to be extracted. See details for available options.

ppmc_interval

The compatibility interval used for determining model fit flags to return (e.g., what = "odds_ratio_flags"). For example, a ppmc_interval of 0.95 will return any PPMCs where the posterior predictive p-value (ppp) is less than 0.025 or greater than 0.975.

quiet

Logical. Should informational summaries and messages be suppressed? Default is FALSE.

Value

The extracted information. The specific structure will vary depending on what is being extracted, but usually the returned object is a tibble with the requested information.

Details

For diagnostic classification models, we can extract the following information:

item_param: The estimated item parameters. This shows the name of the parameter, the class of the parameter, and the estimated value.

strc_param: The estimated structural parameters. This is the base rate of membership in each class. This shows the class pattern and the estimated proportion of respondents in each class.

prior: The priors used when estimating the model.

classes: The possible classes or profile patterns. This will show the class label (i.e., the pattern of proficiency) and the attributes included in each class.

class_prob: The probability that each respondent belongs to class (i.e., has the given pattern of proficiency).

attribute_prob: The proficiency probability for each respondent and attribute.

m2: The M2 fit statistic, including RMSEA and SRMSR indices. Model fit information must first be added to the model using add_fit().

odds_ratio: The observed and posterior predicted odds ratios of each item pair. Model fit information must first be added to the model using add_fit().

odds_ratio_flags: A subset of the PPMC odds ratios where the ppp is outside the specified ppmc_interval.

classification_accuracy: The classification accuracy and consistency for each attribute, using the metrics described by Johnson & Sinharay (2018). Reliability information must first be added to the model using add_reliability().

Methods (by class)

measr_extract(measrdcm): Extract components of an estimated diagnostic classification model.

References

Johnson, M. S., & Sinharay, S. (2018). Measures of agreement to assess attribute-level classification accuracy and consistency for cognitive diagnostic assessments. Journal of Educational Measurement, 55(4), 635-664. doi:10.1111/jedm.12196

Examples

rstn_mdm_lcdm <- measr_dcm(
  data = mdm_data, missing = NA, qmatrix = mdm_qmatrix,
  resp_id = "respondent", item_id = "item", type = "lcdm",
  method = "optim", seed = 63277, backend = "rstan"
)

extract(rstn_mdm_lcdm, "strc_param")
#> Error in extract(rstn_mdm_lcdm, "strc_param"): could not find function "extract"