Class measrfit of models fitted with the measr package

Models fitted with the measr package are represented as a measrfit object, which contains the posterior draws, Stan code, priors, and other relevant information.

Slots

data

The data and Q-matrix used to estimate the model.

type

The type of DCM that was estimated.

prior

A measrprior object containing information on the priors used in the model.

stancode

The model code in Stan language.

method

The method used to fit the model.

algorithm

The name of the algorithm used to fit the model.

backend

The name of the backend used to fit the model.

model

The fitted Stan model. This will object of class rstan::stanfit if backend = "rstan" and CmdStanMCMC if backend = "cmdstanr" was specified when fitting the model.

respondent_estimates

An empty list for adding estimated person parameters after fitting the model.

fit

An empty list for adding model fit information after fitting the model.

criteria

An empty list for adding information criteria after fitting the model.

reliability

An empty list for adding reliability information after fitting the model.

file

Optional name of a file which the model objects was saved to or loaded from.

version

The versions of measr, Stan, rstan and/or cmdstanr that were used to fit the model.

See also