Posterior draws of respondent proficiency

Calculate posterior draws of respondent proficiency. Optionally retain all posterior draws or return only summaries of the distribution for each respondent.

Usage

# S3 method for measrdcm
predict(
  object,
  newdata = NULL,
  resp_id = NULL,
  missing = NA,
  summary = TRUE,
  probs = c(0.025, 0.975),
  ...
)

Arguments

object

An object of class measrdcm. Generated from measr_dcm().

newdata

Optional new data. If not provided, the data used to estimate the model is scored. If provided, newdata should be a data frame with 1 row per respondent and 1 column per item. All items that appear in newdata should appear in the data used to estimate object.

resp_id

Optional. Variable name of a column in newdata that contains respondent identifiers. NULL (the default) indicates that no identifiers are present in the data, and row numbers will be used as identifiers. If newdata is not specified and the data used to estimate the model is scored, the resp_id is taken from the original data.

missing

An R expression specifying how missing data in data is coded (e.g., NA, ".", -99, etc.). The default is NA.

summary

Should summary statistics be returned instead of the raw posterior draws? Only relevant if the model was estimated with method = "mcmc". Default is TRUE.

probs

The percentiles to be computed by the [stats::quantile()] function. Only relevant if the model was estimated with method = "mcmc". Only used if summary is TRUE.

...

Unused.

Value

A list with two elements: class_probabilities and attribute_probabilities.

If summary is FALSE, each element is a tibble with the number of rows equal to the number of draws in object with columns: .chain, .iteration, .draw, the respondent identifier, and one column of probabilities for each of the possible classes.

If summary is TRUE, each element is a tibble with one row per respondent and class or attribute, and columns of the respondent identifier, class

or attribute, mean, and one column for every value specified in probs.