Fit Bayesian diagnostic classification models

Something...

Usage

measr_dcm(
  data,
  resp_id = NULL,
  missing = NA,
  qmatrix,
  type = c("lcdm"),
  method = c("mcmc", "optim"),
  prior = NULL,
  backend = getOption("measr.backend", "rstan"),
  ...
)

Arguments

data

Response data. A data frame with 1 row per respondent and 1 column per item.

resp_id

Optional. Variable name of a column in data that contains respondent identifiers. NULL (the default) indicates that no identifiers are present in the data, and row numbers will be used as identifiers.

missing

An R expression specifying how missing data is coded (e.g., NA, ".", -99, etc.). The default is NA.

qmatrix

The Q-matrix. A data frame with 1 row per item and 1 column per attribute. All cells should be either 0 (item does not measure the attribute) or 1 (item does measure the attribute).

type

Type of DCM to estimate. Currently only the LCDM is supported.

method

Estimation method. Options are "mcmc", which uses Stan's sampling method, or "optim", which uses Stan's optimizer.

prior

A measrprior object. If NULL, default priors are used, as specified by default_dcm_priors().

backend

Character string naming the package to use as the backend for fitting the Stan model. Options are "rstan" (the default) or "cmdstanr". Can be set globally for the current R session via the "measr.backend" option (see options). Details on the rstan and cmdstanr packages are available at https://mc-stan.org/rstan/ and https://mc-stan.org/cmdstanr/, respectively.

...

Additional arguments passed to Stan.

For backend = "rstan", arguments are passed to rstan::sampling() or rstan::optimizing().

For backend = "cmdstanr", arguments are passed to the $sample() or $optimize() methods of the CmdStanModel class.

Value

A measrfit object.